@@ -8,14 +8,13 @@ dynamic = ["version"]
 description = "A polymorphic pattern matching library for Python"
 readme = "README.md"
 license = "MIT"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 authors = [{ name = "linuxdaemon", email = "linuxdaemon.irc@gmail.com" }]
 keywords = ["library", "pattern-matching", "utility", "regex"]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -61,7 +60,7 @@ run-cov = "coverage run -m pytest{env:HATCH_TEST_ARGS:} {args}"
 cov-combine = "coverage combine"
 cov-report = ["coverage report --show-missing --skip-covered", "coverage xml"]
 [[tool.hatch.envs.hatch-test.matrix]]
-python = ["3.13", "3.12", "3.11", "3.10", "3.9"]
+python = ["3.13", "3.12", "3.11", "3.10"]
 
 [tool.hatch.envs.types]
 extra-dependencies = ["mypy>=1.0.0"]
@@ -78,12 +77,12 @@ float_to_top = true
 
 [tool.black]
 line-length = 80
-target-version = ["py39"]
+target-version = ["py310"]
 include = '\.pyi?$'
 
 [tool.ruff]
 line-length = 80
-target-version = 'py39'
+target-version = 'py310'
 
 [tool.ruff.format]
 docstring-code-format = true
@@ -158,7 +157,7 @@ line-length = 120
 
 [tool.mypy]
 namespace_packages = true
-python_version = "3.9"
+python_version = "3.10"
 warn_unused_configs = true
 strict = true
 strict_optional = true
@@ -217,13 +216,18 @@ omit = ["polymatch/_version.py"]
 [tool.commitizen]
 name = "cz_conventional_commits"
 tag_format = "v$version"
-version_scheme = "semver"
+version_scheme = "semver2"
 version_provider = "scm"
 update_changelog_on_bump = true
 major_version_zero = true
+annotated_tag = true
 
 [tool.nitpick]
-style = ["gh://TotallyNotRobots/nitpick/lib-style-3.9.toml"]
+style = ["gh://TotallyNotRobots/nitpick/lib-style-3.10.toml"]
 
 [tool.check-spdx-header]
 headers = ["2018-present linuxdaemon <linuxdaemon.irc@gmail.com>"]
+
+[tool.autoflake]
+remove-all-unused-imports = true
+in-place = true
